Plug.Parsers.XMLRPC v1.0.2 Plug.Parsers.XMLRPC

Parses XMLRPC request body.

An empty request body is parsed as an empty map.

Functions

init(options)

Callback implementation for Plug.Parsers.init/1.

parse(conn, arg2, arg3, headers, opts)

Attempts to parse the connection’s request body given the content-type type, subtype, and its parameters.

The arguments are:

the Plug.Conn connection

type, the content-type type (e.g., "x-sample" for the "x-sample/json" content-type)

subtype, the content-type subtype (e.g., "json" for the "x-sample/json" content-type)

params, the content-type parameters (e.g., %{"foo" => "bar"} for the "text/plain; foo=bar" content-type)

This function should return:

{:ok, body_params, conn} if the parser is able to handle the given content-type; body_params should be a map

{:next, conn} if the next parser should be invoked

{:error, :too_large, conn} if the request goes over the given limit

Callback implementation for Plug.Parsers.parse/5.